Dynamic root page!
The root page now has logic associated to it that is tied in with the a
settings toggle. If toggled (which it automatically is on first open) it
will automatically present the tutorial page. After the tutorial page is
left. the toggle is automatically turned off, however it may be turned
back on by the user.